[FE] firebase methods in folder
